+++ checkpoint/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Validate merged and saved checkpoint params by comparing it against its
  `.zarray` counterpart.
+- Added private utilities to work with array fragments (`_src/arrays`).
 
 ## [0.6.1] - 2024-08-22